recite/requirements.txt
 baidu-aip==2.2.18.0
 PyYAML==5.4.1
 playsound==1.2.2